[T2614] REFACOR: inherit form ebill_postfinance_service.py
