+++ ebill_postfinance_recipient_subscription/models/__init__.py
@@ -1,1 +1,1 @@
-
+from . import ebill_postfinance_service
