+++ ebill_postfinance_recipient_subscription/models/ebill_postfinance_service.py
@@ -1,0 +1,33 @@
+# Copyright 2022 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+import logging
+import logging.config
+
+from ebilling_postfinance import ebilling_postfinance
+
+from odoo import api, fields, models
+from odoo.exceptions import UserError
+
+_logger = logging.getLogger(__name__)
+
+
+class EbillPostfinanceService(models.Model):
+    _inherit = "ebill.postfinance.service"
+
+    def get_ebill_recipient_subscription_status_bulk(self, bill_recipient_id):
+        service = self._get_service()
+        res = service.get_ebill_recipient_subscription_status_bulk(bill_recipient_id)
+        return res
+
+    def initiate_ebill_recipient_subscription(self, recipient_email):
+        service = self._get_service()
+        res = service.initiate_ebill_recipient_subscription(recipient_email)
+        return res
+
+    def confirm_ebill_recipient_subscription(self, initiation_token, activation_code):
+        service = self._get_service()
+        res = service.confirm_ebill_recipient_subscription(
+            initiation_token, activation_code
+        )
+        return res
